Strengthen type of methods accepting thisArg.
Strengthen forEach, filter, map, from, find, findIndex to be more
stricter, so that the TypeScript compiler can understand the type of
this when thisArg is used but the predicate does not explicitly mention
the type of this and only be implicit.

Although methods like some, every also have the thisArg as its parameter
but is not included in this change, because in some places those are
called like e.g. array.some(isNaN) and in these cases, the TSC does not
understand the type of the isNaN is actually a subtype.

Also, this excludes to improve the methods cases where the interface has
overloads of the same method, due to the bug
https://github.com/microsoft/TypeScript/issues/44373, as the current TSC
cannot resolve multiple declarations properly for cases which these
functions are called on a union of types which all are arrays. e.g.
(string[]|number[]).forEach(predicate, ...);